requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.0.0

# ModernBERT dependencies
datasets>=2.13.0
tokenizers>=0.13.0

# Visualization and reporting
matplotlib>=3.7.0
plotly>=5.14.0
tqdm>=4.65.0

# DSPy framework
dspy-ai>=2.0.0
openai>=1.3.0

# Utilities
python-dotenv>=1.0.0
dataclasses>=0.6
jsonschema>=4.17.0

# Optional dependencies for specific functionalities
# Uncomment if needed
# tensorboard>=2.13.0  # For training visualization
# accelerate>=0.20.0   # For faster training
# sentencepiece>=0.1.99  # For specific tokenizers 